+++ src/KardFinancial/Notifications/Types/EmailNotificationPlacementFileData.cs
@@ -1,0 +1,37 @@
+using global::System.Text.Json;
+using global::System.Text.Json.Serialization;
+using KardFinancial.Core;
+
+namespace KardFinancial;
+
+[Serializable]
+public record EmailNotificationPlacementFileData : IJsonOnDeserialized
+{
+    [JsonExtensionData]
+    private readonly IDictionary<string, JsonElement> _extensionData =
+        new Dictionary<string, JsonElement>();
+
+    /// <summary>
+    /// The placement ID, also used as the notification resource ID
+    /// </summary>
+    [JsonPropertyName("id")]
+    public required string Id { get; set; }
+
+    [JsonPropertyName("attributes")]
+    public required EmailNotificationPlacementFileAttributes Attributes { get; set; }
+
+    [JsonPropertyName("relationships")]
+    public required EmailNotificationPlacementFileRelationships Relationships { get; set; }
+
+    [JsonIgnore]
+    public ReadOnlyAdditionalProperties AdditionalProperties { get; private set; } = new();
+
+    void IJsonOnDeserialized.OnDeserialized() =>
+        AdditionalProperties.CopyFromExtensionData(_extensionData);
+
+    /// <inheritdoc />
+    public override string ToString()
+    {
+        return JsonUtils.Serialize(this);
+    }
+}